pkg/obfuscate: fix SQL tokenizer number parsing (#11115)
scanNumber failed to parse certain negative numbers and also failed to
set an error in the tokenizer when it returned LexError, resulting in
nil error messages. This commit causes scanNumber to correctly parse
those numbers and also set a more meaningful error message when it
encounters an error.

This also adds some tests specifically for the number parser.