pkg/trace: Support for dynamic sampling rates and rare sampler in AgentPayloads (#14637)
When we send an AgentPayload to the backend, the TargetTPS, ErrorTPS and RareSamplerEnabled now reflect the actual configuration of the samplers instead of the values coming from the static configuration. This will allow the backend to know the effective sampler configuration for each agent. See the pull request for more information.