[Windows] implement mapping of pid to service name
Checks the pid against the table of SCM controlled processes.
If it's SCM controlled, returns the service information.

Because we must enumerate the entire SCM (there doesn't seem to be an api
for that), the SCMManager object maintains a cache of objects, and refreshes
only when it sees a PID it hasn't seen before.

On a machine with high process churn, this could still result in a lot
of accesses.  However, if process agent queries only when doing the
process check (i.e. every 30s), then it should only iterate the list
once per 30s.