epforwarder: double dbm-samples defaultBatchMaxContentSize
We have been seeing some issues where data fails to send
to our dbm samples endpoint. We see far fewer examples of
the dbm-activity and dbm-metrics pipelines being blocked or
failing to send data, and we suspect that this is partly
due to the drastic differences in size between those
pipelines and the dbm-samples pipeline.

This change doubles the defaultBatchMaxContentSize for
the dbm-samples pipeline to try to address this.