[single-machine-performance] Push agent containers to SMP ECR (#14438)
* [single-machine-performance] Push agent container to SMP ECR

This commit is an attempt to introduce pushing containers from Agent CI for
single-machine-performance's Regression Detector in our isolated
infrastructure. Much like we have done for vectordotdev/vector we intend to run
the Regression Detector on Agent changes, giving a reasonable statistical
guarantee that a change does or does not modify Agent performance by more than
random chance. In order for the Regression Detector to run jobs it must have
access to a 'baseline' and 'comparison' target. Baseline in this project would
be a container built from current `main` branch, comparison would be a container
built from the tip of a PR.

The main thing demonstrated here is that the team credentials SMP has created
for Agent are functional and are able to push up a containers, in a way that is
acceptable to Agent Platform. I have ammended `.docker_build_job_definition` to
mirror every created container to single-machine-performance's ECR, noting that
the tag now avoids the use of `CI_PIPELINE_ID`. In a later commit we will
introduce job submission and will rely on being able to compute the tag of a
previous pipeline's container from available Gitlab metadata, specificall
`CI_COMMIT_SHA` for the comparison container and whatever metadata maps to the
base branch's current SHA, `CI_MERGE_REQUEST_SOURCE_BRANCH_SHA`?

There are two outstanding questions regarding this work that I am aware of:

* Is there a race condition present between the triggering of this pipeline vs
main if users squash commits?
* Should we grant the exisitng CI user permissions into
single-machine-performance rather than use an issued bot account as done
presently and for vectordotdev/vector?

We've successfully demonstrated pushing up containers in a previous iteration of
this work, see https://gitlab.ddbuild.io/DataDog/datadog-agent/-/jobs/195939127.

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>

* PR feedback

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>

* trim ECR URL out of destination

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>

* correct job dependency

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>

* drop parallel.matrix

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>