Adding 'cloud_provider_source" to the inventories payload
We now track from where we fetch metadata related to a cloud provider.
This only support AWS EC2 for now. Depending on the instances
types, configuration, ... the Agent can use multiple sources to deduce
it's running one EC2.
The source used is now sent as 'cloud_provider_source' in inventories.