Avoid race condition when running in serverless mode.
When run in AWS Lambda, the Datadog Lambda Extension will start both the
otlp pipeline and the trace agent in the same process. This means that
there is both a grpc client and server running at the same time. In both
cases, grpc attempts to load and use its logger. However, this logger is
a global instance, and thus the same logger instance is used for both
the otlp pipeline client and the trace agent server. This logger lacks
proper locking.

Therefore, to avoid race conditions when attempting to use this singular
logger between both the client and server, it is best to disable logging
here in the client. Now, only the server will be using the logger and
thus the race conditions will be avoided.