[workloadmeta] Log only bundle size on ack timeout
When a workloadmeta subscriber cannot close the EventBundle's channel in
time, the store logs a warning message that it will move on to the next
subscribers even though the one being notified at the moment did not
acknowledge in time.

Originally, we used to log the entire contents of the bundle: all of the
events and their associated entities, and the bundle's source. Several
refactorings later, bundles no longer have a source, and the events now
only contain a pointer to the entity, which makes the log message just a
list of pointers, which is not very useful for debugging. Additionally,
with the introduction of a cluster-level workloadmeta, the number of
entities has grown very large, making these errors consume a lot of
space.

This warning is still useful information though, so now instead of
dumping the whole bundle, we now just show its size.