fix(RCM): Fix timeout of client bypass
Before this commit, when the previous request was still pending,
the client bypass wouldn't timeout until it managed to send its
request, meaning that the effective maximum timeout for a new client
was (request TTL) + (bypass TTL).

This was made more visible when the refresh interval went from 60s
to 5s minimum.

Now, the client bypass timeout takes the previous request into account
as well, so that a client doesn't wait for more than 2s on its first
request