[system-probe] make sure stats are reported as soon a module registration happens
When modules are registered in the system-probe, a process to update
stats in the global `l` field every 15 seconds starts.

Those stats are pulled in by the `agent status` command which user can
use to see what's running in the agent.

The loop was written as:

```
for now := <-ticker {
  // update stuff
}
```

Which mean that for the first 15 seconds fo the agents life, stats would
come back incomplete. Meaning that for the first 15 seconds of the agent's.

This PR changes the stats loop to compute stats *then* wait for 15
seconds.

QA instructions (meant to run on a vanilla linux host):

* in one terminal window run `watch "sudo -u dd-agent
  /opt/datadog-agent/bin/agent/agent  status  | head -n10"`
* in another terminal window, restart the system-probe
* you should *not*  see errors about systemprobe.tmpl failing to render.
  Running these steps on 7.42 and below shows:

  ```
  Getting the status from the agent.
  template: /systemprobe.tmpl:11:34: executing "/systemprobe.tmpl" at <.updated_at>: invalid value; expected float64
  ```

  For the first 15 seconds.

  You should *not see* see these errors in the first 15 seconds of the
  system probe running (or ever).