[RCM] Seed the root version as 1 when using an unverfied TUF store
The backend expects the agent and all of its clients to have version 1
of the root file to kick off the root update process. Even if we aren't
verifying the root chain yet in all clients, we need to pretend as if we
have the 1st root to pass the backend checks.