[Serverless] Enable OTLP endpoint in extension. (#15321)
* Enable OTLP endpoints when running in serverless.

* Add unit tests for ensuring collector is running and accepting traces.

* Address race condition in test.

* Avoid race condition when running in serverless mode.

When run in AWS Lambda, the Datadog Lambda Extension will start both the
otlp pipeline and the trace agent in the same process. This means that
there is both a grpc client and server running at the same time. In both
cases, grpc attempts to load and use its logger. However, this logger is
a global instance, and thus the same logger instance is used for both
the otlp pipeline client and the trace agent server. This logger lacks
proper locking.

Therefore, to avoid race conditions when attempting to use this singular
logger between both the client and server, it is best to disable logging
here in the client. Now, only the server will be using the logger and
thus the race conditions will be avoided.

* Avoid race conditions by calling under lock.

* Add build tags and methods for with/without otlp.

* Add dependencies from otlp serverless testing.

* Always start trace and metric agents on a free port to avoid collisions.

* Ensure internal otlp trace endpoint uses different port.

* Wait for collector to fully shut down.

* Update licenses for enabling otlp in serverless.

* Use the returned value from waitForState.

* Set hostname during test to prevent hostname resolution errors.

* Serverless otlp release note.

* Pair wg.Add with each goroutine.

* Include comment for new config setting.

* Update releasenotes/notes/serverless-enable-otlp-endpoint-6134c9fc34254599.yaml

Co-authored-by: Rosa Trieu <107086888+rtrieu@users.noreply.github.com>

* Separate building and starting pipeline.

* Pull state strings from otelcol.State.

* Without otlp buildtag.

* Specify package.

* Do not return error that will not be checked.

* Readd returning error to pipeline.Run.

* Docstrings.

* Docstrings.

* Docstrings.

---------

Co-authored-by: Rosa Trieu <107086888+rtrieu@users.noreply.github.com>
Co-authored-by: Ivan Topolcic <ivan.topolcic@datadoghq.com>