[RCM] Prevent warning message from showing when no refresh_interval set (#15549)
We previously made a change to eliminate the default refresh interval in
the config module in order to be able to detect when no interval was
provided so we could safely allow the RC backend to set the value
appropriately. Unfortuantely, this caused a warning message to be
displayed on agent startup when remote config was enabled, as "" is not
a valid duration string.

To prevent this, we first check if the value is set. If it isn't, we
don't attempt to parse the empty string as a duration and just use the
default refresh interval.