[single-machine-performance] file_to_blackhole: test egress throughput
This pull request intends to be a follow-up to PR #14528. Currently,
regression detector jobs for the `file_to_blackhole` experiment use
ingress throughput as an optimization goal. Recent work on
single-machine-performance now enables users to set a single
optimization goal per experiment instead of a single optimization goal
for all experiments.

This pull request changes the optimization goal of the
`file_to_blackhole` experiment to egress throughput, which is of more
interest than ingress throughput for that experiment. In addition,
this commit updates the Regression Detection README to include
information on how to set an optimization goal for each experiment.

If an `experiment.yaml` file is absent for an experiment, the
optimization goal for that experiment defaults to ingress
throughput. This value can be changed for an experiment by placing an
`experiment.yaml` file at the root of the directory tree for a given
experiment, e.g., for the `file_to_blackhole` experiment, at
`test/regression/cases/file_to_blackhole/experiment.yaml`.

The contents of an example `experiment.yaml` file are

```yaml
optimization_goal: egress_throughput
```

As the default, `ingress_throughput` is currently the only other
supported value of `optimization_goal`.

Signed-off-by: Geoffrey M. Oxberry <geoffrey.oxberry@datadoghq.com>