Dogstatsd server component scaffolding (#15283)
This change introduces the component interface for dogstatsd - starting with the dogstatsd server which is the public API the rest of the agent consumes.

This is part 1 of 2. In part 1 we:

Provide a component interface for the dogstatsd server (a shim that proxy's it's calls to the original implementation)
Introduces a bundle for all dogstatsd components (there will be several)
Add the dogstatsd component dependency to existing FX apps
Small refactors to decouple dogstatsd from other parts of the agent code:
moving the global DSD reference out of cmd/agent/common
Removing global variables where possible (when not blocked by other refactors)
add a Start method to dogstatsd to separate initialization from server startup.
Part 2 will:

Migrate pkg/dogstatsd/server.go, tests, and related pieces into the component implementation.
After part 2 we can start breaking apart sub-dependencies of the dogstatsd server into components.