test/regression: update lading & smp versions
To enable the `egress_throughput` change to `file_to_blackhole`, the
versions of `smp` and `lading` must be updated. Actually, another
`smp` release is needed and the `smp` version used by `datadog-agent`
must be updated to *that* version, but the version updates in this
commit are the most recent versions available at the moment.

Signed-off-by: Geoffrey M. Oxberry <geoffrey.oxberry@datadoghq.com>