[single-machine-performance] file_to_blackhole: change optimization goal to egress throughput (#15631)
* [single-machine-performance] file_to_blackhole: test egress throughput

This pull request intends to be a follow-up to PR #14528. Currently,
regression detector jobs for the `file_to_blackhole` experiment use
ingress throughput as an optimization goal. Recent work on
single-machine-performance now enables users to set a single
optimization goal per experiment instead of a single optimization goal
for all experiments.

This pull request changes the optimization goal of the
`file_to_blackhole` experiment to egress throughput, which is of more
interest than ingress throughput for that experiment. In addition,
this commit updates the Regression Detection README to include
information on how to set an optimization goal for each experiment.

If an `experiment.yaml` file is absent for an experiment, the
optimization goal for that experiment defaults to ingress
throughput. This value can be changed for an experiment by placing an
`experiment.yaml` file at the root of the directory tree for a given
experiment, e.g., for the `file_to_blackhole` experiment, at
`test/regression/cases/file_to_blackhole/experiment.yaml`.

The contents of an example `experiment.yaml` file are

```yaml
optimization_goal: egress_throughput
```

As the default, `ingress_throughput` is currently the only other
supported value of `optimization_goal`.

Signed-off-by: Geoffrey M. Oxberry <geoffrey.oxberry@datadoghq.com>

* test/regression: update lading & smp versions

To enable the `egress_throughput` change to `file_to_blackhole`, the
versions of `smp` and `lading` must be updated. Actually, another
`smp` release is needed and the `smp` version used by `datadog-agent`
must be updated to *that* version, but the version updates in this
commit are the most recent versions available at the moment.

Signed-off-by: Geoffrey M. Oxberry <geoffrey.oxberry@datadoghq.com>

* regression detector: remove obsolete flag

The `--target-inherit-environment` flag in `smp` has been removed
because targets now inherit environment variables by default, as of
`smp` 0.7.0. This commit removes that obsolete flag from `smp` calls
in the Agent CI configuration.

Signed-off-by: Geoffrey M. Oxberry <geoffrey.oxberry@datadoghq.com>

* test/regression: update smp version to 0.7.1

To enable per-experiment optimization goal support, the version of
`smp` used by Agent must be version 0.7.1 or later. This commit
updates the `smp` version used by Agent's CI pipeline to version 0.7.1
from 0.7.0 for that support.

Signed-off-by: Geoffrey M. Oxberry <geoffrey.oxberry@datadoghq.com>

---------

Signed-off-by: Geoffrey M. Oxberry <geoffrey.oxberry@datadoghq.com>