Correct overly optimistic error range
Given that the sliding window algorithm runs in a separate goroutine,
we can not assume any number of iterations to happen before the end of
the test. Lowest was no polls at all. With previous commit, we can
expect zero utilization to be reported in this case. Upper boundary is
harder to estimate, but max of 50k runs was 0.49.