Fixes incorrect err code from compressor.AddItem (#15613)
* Fixes incorrect err code from AddItem

This change ensures AddItem will return correct err code in edge
cases where len(data) is less than "maxZippedItemSize" but greater
than "maxPayloadSize".

"maxZippedItemSize" is calculated from the passed argument
"maxUncompressedSize" whereas "maxPayloadSize" is a separate argument.
'checkItemSize' must consider both in order for AddItem to correctly return
'ErrItemTooBig' instead of 'ErrPayloadFull' in these certain cases.

Without this change, the failure case is when a compressor is brand-new,
ie has an empty payload. When AddItem is invoked with payloads of
certain lengths, the first check to 'checkItemSize' would _pass_
(incorrectly) and then it would later fail after a call to
`checkItemSize`. This would case 'ErrPayloadFull' to be returned instead
of 'ErrItemTooBig'

* Update pkg/util/compression/zlib.go

Co-authored-by: Srdjan Grubor <sgnn7@sgnn7.org>

---------

Co-authored-by: Srdjan Grubor <sgnn7@sgnn7.org>