Fix race condition which returned partial metric samples.
WaitForSamples is written so that it returns once _any_ sample is
received. This can be a problem if you are waiting on multiple samples.
In which case, it is possible that only _some_ of the expected samples
are received by the demultiplexer when the ticker fires, then seeing
that a sample was received, returns this partial slice.

Instead, WaitForNumberOfSamples allows us to continue waiting until at
least a given number of samples are received, thus avoiding this race
condition.