Introduce UDS DogStatsD regression experiment
This commit adds an experiment for the Regression Detector that runs UDS
DogStatsD load into the Agent. This is one of the more common configurations of
the Agent. The approach taken is similar to our other RD runs. I have updated
lading to a version that includes self-throttling based on target
performance. This version of lading will search for Agent's stable throughput
capacity and push load at that level. The previous version of lading pushed as
hard as possible, causing chaotic results from Agent.

See also #15614 and #15613.

REF SMP-371

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>