Re-write utilization tracker to avoid race condition in tests
Sliding window polling goroutine was not synchronized with the test
execution (which is advancing time on the mock clock), and, depending
on how goroutine were scheduled, could run poll zero, one or arbitrary
number of times during the test, making the results
unpredictable. Also it could poll twice in a short succession, causing
poll function to divide by zero and produce NaN utilization.

This patch rewrites utilization tracker to operate synchronously
internally, communicating inputs and outputs via channels. Time is
sampled on the main goroutine only. The results are still
unpredictable insofar goroutine scheduling is involved, but in tests
it can be controlled by carefully placing channel reads.

It also shifts code around a bit, separating expvar and ticker
interaction from the core tracker code to make it more testable.

Finally, the smoothing method is changed from sliding window to a
exponential weighted average, for the only reason that it makes average
move more predictably towards the signal, which also helps with
testing.