Refactor path handling
Construct capture file path, validate it and open the real file
synchronously in the TrafficCapture#Start, instead of doing this in
background in Writer#Capture and then waiting for it.

This removes the need for the writer to store file object and the
path, and to synchronize access to them. We can also get rid of the
global filesystem object and its associated locks, as all fs related
operations are contained inside validateLocation and OpenFile.