Fix a race in the dogstatsd capture writer (#15802)
* Refactor path handling

Construct capture file path, validate it and open the real file
synchronously in the TrafficCapture#Start, instead of doing this in
background in Writer#Capture and then waiting for it.

This removes the need for the writer to store file object and the
path, and to synchronize access to them. We can also get rid of the
global filesystem object and its associated locks, as all fs related
operations are contained inside validateLocation and OpenFile.

* Stop the timer goroutine if capture was stopped early

* Fix race condition in the capture writer

Capture did not completely drain Traffic channel after StopCapture was
called, causing TestWriter* tests to fail on mismatch of sent vs read
payloads.

This patch re-arranges channel handling and locking to make sure that
if Enqueue returns true for a payload, it will be written to the
file (modulo io errors).

Several writer methods that were using a lock are only called from the
main goroutine and are not accessing shared data concurrently. These
were marked private, and locking was removed.

Shutdown channel was removed, and stopped capture is signalled to the
main writer goroutine via closure of the Traffic channel.

RWMutex that was part of the writer is repurposed to synchronize
access to ongoing, accepting and Traffic, the only fields that are
accessed concurrently.