Mock events and service check channels
TestUDPReceive uses mock demultiplexer to receive service
checks (among others). Before this patch, the channel that the test
was reading service checks from, was also connected to a running
aggregator instance, that would once in a while read the item from
the channel before the test, casuing the test to fail.

This patch adds a new method to the DemultiplexerWithAggregator
interface that abstracts over how the channels are stored, and mock
overrides this method to provide mocked channels for both the test and
the dogstatsd server.