[workloadmeta/containerd] Fix race condition in SBOM collection (#15834)
* [workloadmeta/containerd] Fix race condition in SBOM collection

Fixes a race condition that makes the sbom extraction fail sometimes because it
cannot find the image in workloadmeta. The reason is that when we notified
workloadmeta of a new image, we immediately enqueued the sbom collection job
without waiting until workloadmeta had time to process the event and store the
image.