[pkg/collector/python] Force usage of stdlib's distutils in Python (#16024)
Sets the SETUPTOOLS_USE_DISTUTILS environment variable to stdlib before starting the embedded Python interpreter.

Alternative fix to #15960, without removing setuptools from the Agent, which is causing issues when installing some integrations / dependencies in the embedded Python environment. Will be accompanied with a revert of #15960: #16025.

Using the default value (local) with setuptools >= 60 present in the Python environment would cause any import of distutils to load setuptools' embedded _distutils (see setuptools' changelog), which causes a 10MB usage increase in the Agent.
distutils is imported in the network check, which runs by default, so that 10MB usage increase would be present on all Agent deployments.