usm: http2: differentiate between header filterting and header processing
We used to have a single constant to define how much headers we filter and process,
but in the reality there is a difference. We have to filter as much as possible headers
but we are looking for 2 (soon to be 3) headers in the requests (path, method, and content-type in the future)
and in response we look for a single header (status).
The change reduces the binary size by 25% compared to base branch.