Add "container_include_condition" option to config.datadog
Before this commit we could only use:

 * container_include
 * container_exclude

To choose which containers will be monitored by datadog.
And container_include takes precedence over container_exclude.

This makes it very hard/impossible to select which containers should
be monitored in a fine-grained manner.

In addition, if container_include is set it might ignore pause containers
which are supplied by the agent when exclude_pause_container is set as
detailed in the following issue:

  https://github.com/DataDog/datadog-agent/issues/6599

This commit adds `container_include_condition`, which lets you use
conditional expressions to better specify which containers to monitor.

More information about CEL, the conditional expression language used:

  https://github.com/google/cel-spec
  https://github.com/google/cel-go

For example, you can use the following condition expressions in the
configuration:

  ```
  name == 'eu_gcr' || name == 'private_jfrog'
  name == 'secret-container-dd' && image == 'docker-dd-agent'
  name.matches('secret-container-.*') && image == 'docker-dd-agent'
  ```

The `container_include_condition` option is compatible with the
`exclude_pause_container` option.