remoteconfig: guard against empty updates causing repository errors
This is primarily a defensive change for any upstream clients using the
`Repository` structure in `pkg/remoteconfig`. Because the agent sends an
`{}` payload to tracers when there isn't an update, if a client directly
tries to serilaize this into an `Update` field it could result in the
client calling Update with an empty payload. This previously caused a
JSON parsing error, which registers as a client update error and is
reported to the backend despite this being a perfectly normal situation.