Calling cloud providers in parallel to detect host aliases
As we add more cloud provider the host aliases detection now take up 10
to 15s in the worst case. This blocks the agent startup and is
particularly visible on Windows when using the GUI (since the GUI wait 10
to 15s to get the status page).

Calling each cloud providers in parallel should reduce the delay to the
slowest provider.