pkg/trace/api: Deprecate and remove usage of UsePreviewHostnameLogic.
Follow up to: [PR](https://github.com/DataDog/opentelemetry-mapping-go/pull/21).

This PR Deprecates field `UsePreviewHostnameLogic` and removes usage of this field. This is in an effort to graduate the `exporter.datadog.hostname.preview` feature gate to stable ([task](https://datadoghq.atlassian.net/browse/OTEL-231)).

Once this change is released, the next steps are:
- opentelemetry-collector-contrib: Mark the `exporter.datadog.hostname.preview` feature gate as stable. Replace deprecated functions calls by new function calls (following [PR](https://github.com/DataDog/opentelemetry-mapping-go/pull/21)). Remove references of `UsePreviewHostnameLogic` field (e.g. [here](https://github.com/open-telemetry/opentelemetry-collector-contrib/blob/v0.73.0/exporter/datadogexporter/traces_exporter.go#L179)).
- opentelemetry-mapping-go: remove deprecated functions and codepaths which use preview rules.