Make uriPasswordReplacer less aggressive (#15959)
* Cleanup uri password regexp

Remove escapes from symbols that do not need escaping.
Set case-insensitive flag to simplify character ranges.

* Make uriPasswordReplacer less aggressive

Add `|` and `"` to the list of characters that the password field can
not contain. While these may be present in a naively constructed URL
string, these characters are not valid according to the spec. URLs
that contain these characters are rejected by Go `net/url` and
everything that uses it.