reduce http2 complexity (#16155)
* usm: http2: differentiate between header filterting and header processing

We used to have a single constant to define how much headers we filter and process,
but in the reality there is a difference. We have to filter as much as possible headers
but we are looking for 2 (soon to be 3) headers in the requests (path, method, and content-type in the future)
and in response we look for a single header (status).
The change reduces the binary size by 25% compared to base branch.

* usm: http2: Simplying a condition and removing redundant const

The const HTTP2_MAX_FRAME_TYPE is redundant as we can use kContinuationFrame, which is the header
type with the highest value.

* usm: http2: Fetching http stream only once

The change spares 2% of the original code size. Furthermore, it access fewer times to the
inflight maps which reduces the runtime

* usm: http2: removing redundant parameters

This change gives us a reduction of 3% in instructions

* usm: http2: stop using _with_telemetry functions.

The change saves 4.5% of the instructions

* usm: http2: Read buffer when needed instead of cache

Instead of reading 160 bytes for each headers frame, we read as much we need.
This simplifies the code, as we don't need to work with buffer and offset,
and it allows us to better handle long frames (more than 160) bytes.

Reducdes 6% from the original binary.

* usm: http2: copied read_into_buffer_

* usm: http2: changed global dynamic counter access

Instead of fetching the global counter from a map with every iteration, we do that before the loop
and we get it as a pointer. Saving 4.4% from binary size

* usm: http2: Save dynamic value only if relevant

Moved the code to save dynamic value to the interesting headers loop.
It increases the binary size by 3%, but fixes a bug introduced in the code.

* usmg: http2: using static functions to determine if a static index is interesting

Instead of accessing the maps multiple times, we are using a static functions
to determine if the index belongs with the statis table or not.
Saved 1.33% from binary compared to the last commit

* usm: http2: Sparing a redundant call for read_var_int

Reorganizing the code so we can remove a call for read_var_int.
Saves 4.7%

* usm: http2: Simplying parse_field_literal to reduce complexity

Rearranging the if-clauses so it spares 0.5%

* usm: http2: Simplying filter_relevant_headers to reduce complexity

Rearranging the if-clauses so it spares 2%

* usmg: http2: Simplifying the main loop

Instead of checking 2 different boundaries in the loop, we're checking the
minimal boundary of them. Although it is very unlikely that a frame will
exceed the packet limit, but we want to make sure we're dealing that case,
and we must check the skb->len for the verifier.

Reducing 0.5%

* usm: http2: rearrnging loops to simplify the code

Reducing code size by 11%

* usm: http2: simplifying kernel of http2

Got rid of parse_result and converted it into a boolean. Move to usermode the conversion of status code and method.
Simplify if clauses. Sparing total of 1.4% of the binary size

* usm: http2: Removed old commit files

* Fixed cr comments