usm: javatls: test cleanup and fixing flakiness
The Java injection tests were flaky because the inject JAR first
wrote a log and then created a file. We waited for the Docker
logs to show the log, but as soon as the log was introduced,
the host Docker was terminated. Therefore, there was a race
condition in which the JAR did not create a test file that was
expected to be created.