clusteragent: distinguish data age and uptime date
The data age might be significantly in the past even for fresh data,
in particular when using a large `rollup(<duration>)`. This makes it
impractical to use because you intially need to wait up to `maxAge` to
see your metric being updated.

This change distinguishes the two things. `DataAge` is still being used
for the `update time` condition to avoid breaking the printer behavior, but
the plan is to eventually introduce a new field for the data age.