Starts splitting config helpers from the config package
As we continue to migrate the Agent code to components the many helpers
in pkg/config need to be moved somewhere else. The goal is to have a
simple API for the Config component while offering helpers on top of it.

The goal is to have helpers that can take either a "pkg/config:Config" or a
"comp/core/config:Component". This will facilitate the migration to
components.

For now the helpers are moved to pkg/config/utils. This is temporary as
we brainstorm the best place for them. Starting to split them from the
main config package is still useful.