regression_detector: get pr-commenter the hard way
The `/usr/local/bin/pr-commenter` command doesn't exist in the Docker
image used to run the regression detector commands, so we have two
options:

* build another image with `pr-commenter` in it

* install `pr-commenter` as part of the CI job

The former is more elegant, but the latter is easier to stand up and
test, so this commit takes the latter approach. Rather than chain
together a lot of commands on one line, this commit splits commands by
line in the possibly vain hopes that the result will be easier to
debug than the verbose one-liner.

Signed-off-by: Geoffrey M. Oxberry <geoffrey.oxberry@datadoghq.com>