[single-machine-performance] post report to GitHub (#16500)
* [single-machine-performance] post report to GitHub

This commit & pull request emits an HTML report from the Single
Machine Performance regression detector to a GitHub PR comment using
the [GitHub PR Commenter
Bot](https://github.com/DataDog/devtools/tree/master/deb/pr-commenter)
tooling from the DataDog devtools repo, by analogy to other uses of
the GitHub PR Commenter Bot in Agent's GitLab CI configuration.

Signed-off-by: Geoffrey M. Oxberry <geoffrey.oxberry@datadoghq.com>

* regression_detector: get pr-commenter the hard way

The `/usr/local/bin/pr-commenter` command doesn't exist in the Docker
image used to run the regression detector commands, so we have two
options:

* build another image with `pr-commenter` in it

* install `pr-commenter` as part of the CI job

The former is more elegant, but the latter is easier to stand up and
test, so this commit takes the latter approach. Rather than chain
together a lot of commands on one line, this commit splits commands by
line in the possibly vain hopes that the result will be easier to
debug than the verbose one-liner.

Signed-off-by: Geoffrey M. Oxberry <geoffrey.oxberry@datadoghq.com>

* smp pr-commenter: add missing `--package` flag

I made a syntax error when I added a line to invoke `dd-package` when
I neglected to insert a `--package` flag between `binaries.ddbuild.io`
and `devtools/pr-commenter`. This commit rectifies that error by
inserting the missing flag.

Signed-off-by: Geoffrey M. Oxberry <geoffrey.oxberry@datadoghq.com>

* smp ci: try dd-package without --distribution flag

The `--distribution "20.04"` portion of the `dd-package` command seems
to be ignored, so this commit elides it to see what happens.

Signed-off-by: Geoffrey M. Oxberry <geoffrey.oxberry@datadoghq.com>

* smp ci: install wheel to install pr-commenter

Without installing `wheel`, the `dd-package --package
devtools/pr-commenter` command will fail due to an `invalid command
'bdist_wheel'` error. The debug loop for CI is so long that it's
probably better to install slightly more than I think I'll need so
that I can move past the `pr-commenter` installation blocker.

Signed-off-by: Geoffrey M. Oxberry <geoffrey.oxberry@datadoghq.com>

* regression_detector: correct my lousy pip syntax

My `pip` command syntax was buggy, so this commit fixes the incorrect
version specifier and also corrects the version of `setuptools`
installed.

Signed-off-by: Geoffrey M. Oxberry <geoffrey.oxberry@datadoghq.com>

* regression_detector: install dd-package-dev first

Maybe `dd-package` doesn't do dependency resolution? If so, then
`devtools/dd-package-dev` needs to be installed before
`devtools/pr-commenter`, so this commit adds commands to install the
former before installing the latter.

Signed-off-by: Geoffrey M. Oxberry <geoffrey.oxberry@datadoghq.com>

* regression_detector pr-commenter: omit --for-repo

The `pr-commenter` error message shown in CI seems to indicate that
the `--for-repo` flag is invalid (error message was `pr-commenter:
error: unrecognized arguments: --for-repo=datadog-agent`) so this
commit removes that flag.

Signed-off-by: Geoffrey M. Oxberry <geoffrey.oxberry@datadoghq.com>

* regression_detector.yml: clean up pipeline

While going through the process of debugging `pr-commenter` errors, I
accumulated some clutter in the `regression_detector.yml` that can
safely be deleted. This commit deletes that clutter, and also adds a
notification message to indicate that the `bdist_wheel` error during
`pr-commenter` installation is a known harmless error.

Signed-off-by: Geoffrey M. Oxberry <geoffrey.oxberry@datadoghq.com>

* regression_detector.yml: make warning block nicer

GitLab echoes every command, which is problematic when trying to
create a text block to use as a warning or notice. One way to silence
this echo behavior is to use quoted strings; this commit takes that
approach in order to print a text block.

Signed-off-by: Geoffrey M. Oxberry <geoffrey.oxberry@datadoghq.com>

---------

Signed-off-by: Geoffrey M. Oxberry <geoffrey.oxberry@datadoghq.com>