Adding params to the flare comp
The flare component is no longer a passthrough to the pkg/flare but
holds all the logic to create and send flares.

The API for the component has been cleaned and params are now given at
initialization instead.

The logic to send flare is, for now, in the helpers package to be used
by binaries still relying on pkg/flare.