Optional limit on number of dogstatsd contexts per origin
Add a configurable limit on the number of contexts that each dogstatsd
sender is allowed to create in the agent.

For simplicity, in configurations with multiple parallel time sampler
pipelines, the limit is equally divided between the pipelines. This
may make the effective limit slightly smaller due to uneven
distribution of contexts between the samplers (around 5% based on
testing), but the upside is very simple concurrency free
implementation of the limiter.