[RCM] cluster-agent: initialize remote-config after cluster id creation
The remote config client in the cluster-agent needs to report the
Datadog created cluster ID to the backend to help with config routing.
This is established by the cluster-agent. Remote config was previously
being initialized before this call is made, it needs to be initialized
afterwards so that if it doesn't exist it is created.