Add experimental DogStatsD dropped-data checks
This adds a new experiment for DogStatsD over UDP. Both this and the
existing UDS check have an experimental check to verify that the
number of bytes reported by `agent` roughly matches the number of bytes
that was sent by the load generator.

These checks will not impact regression pass/fail status.