Clean up telemetry for context limiter
Unify dropped samples metric the tags limiter, and give it slightly
more descriptive name.

Acceptedd samples metric is probably not useful, and could be
confusing as it only counts samples that create new contexts. Compared
with drops it will always be much lower, which may give impression
that the agent is not handling most of the samples.