Add a status page verbose mode with per-tailer logs stats (#16814)
This PR adds the ground work to easily display per-tailer stats on the status page when agent status -v is used.

The status page is getting very large, but there is still demand to expose more diagnostics information. This PR adds some needed diagnostics information for auto multi line detection as well as introduces some new constructs to make this easier in the future.