Do not add otlp batch processors when running in serverless.
In serverless, we need to be able to flush all data at the end of each
invocation because it is not guaranteed that there will be another
invocation before the container shuts down.  It is okay however for the
data to finish completely flushing when the container fully shuts down.

The batch processor gets in the way of this forced flushing.  Even when
shutting down the collector explicitly, it is still possible for data to
remain "stuck" in the batch processor.