Fix some misc test cleanup (#15807)
* Fix some test cleanup

* Use t.TempDir() instead

* [RCM] Add a Stop method to Service to allow for clean-up in tests (#15877)

* [RCM] Add a Stop method to Service to allow for clean-up in tests

The bolt DB is designed to be persistent for the RC service, but we
don't want it persisting between test runs. A PR made to attempt to
clean up the DB file creating during tests was failing as there was no
way to delete the DB due to the fact the instantiated test Service
instance still had a handle to it. In order to allow the test code to
clean up the test bolt DB, we need a way to shut down the Service and
tell it to close the DB. The new Stop method leverages the
context.CancelFunc we create when starting the main goroutine to both
stop the main goroutine and then shutdown the bolt DB to allow test
cleanup to proceed.

* Use testing.Cleanup() to clean up test code

Less code is best code

* empty commit

---------

Co-authored-by: Kyle Ames <kyle.ames@datadoghq.com>