Optional limit on number of dogstatsd contexts per origin (#16734)
* Optional limit on number of dogstatsd contexts per origin

Add a configurable limit on the number of contexts that each dogstatsd
sender is allowed to create in the agent.

For simplicity, in configurations with multiple parallel time sampler
pipelines, the limit is equally divided between the pipelines. This
may make the effective limit slightly smaller due to uneven
distribution of contexts between the samplers (around 5% based on
testing), but the upside is very simple concurrency free
implementation of the limiter.

* Always add limiter key to the telemetry tags

* Clean up telemetry for context limiter

Unify dropped samples metric the tags limiter, and give it slightly
more descriptive name.

Acceptedd samples metric is probably not useful, and could be
confusing as it only counts samples that create new contexts. Compared
with drops it will always be much lower, which may give impression
that the agent is not handling most of the samples.

* Update tests

* Rename and add comments for clarity

* Clarify behavior when the key is not present on a context.

* Remove debugging Println

* Add comment about the default value for the contexts limit.

* Add more comments

* Rename for clarity