Update root certificates before tests and builds (#17207)
Updates the Windows CI tests and builds scripts to force the update of root certificates used in the image before they are used to download dependencies (as part of the MSI tests and builds).

There is a possible timing issue in the test and build scripts: the base container doesn't contain some of the root certificates needed to install dependencies. These certificates are added while the container is running (probably by a Windows Update process), but if this happens too late, the test and build Windows jobs can fail.