Optional limit on number of dogstatsd tags (#16857)
* Add optional limit on the number of tags stored

In situations with limited resources we don't want to spend them on
storing metrics that would be dropped by the intake. To minimize
impact, we apply the limit to the tags as they would be sent out,
after origin detection and deduplication. The only place where the
tags exist in this form and before we committed to keeping the metric
is in the context resolver.

* Update pkg/config/config.go

Co-authored-by: Scott Opell <me@scottopell.com>

---------

Co-authored-by: Scott Opell <me@scottopell.com>