single-machine-performance: always upload metadata
The existing `single-machine-performance-regression_detector` job
currently only uploads its metadata, in the form of text file called
`submission_metadata`, on job success because GitLab CI's [default
setting for the `artifacts:when` field is
`on_success`](https://docs.gitlab.com/ee/ci/yaml/#artifactswhen). This
metadata is also helpful for debugging purposes when jobs fail, so
this commit changes the `artifacts:wnen` setting in the
`single-machine-performance-regression_detector` job to `always` so
that this metadata is uploaded unconditionally.

Signed-off-by: Geoffrey M. Oxberry <geoffrey.oxberry@datadoghq.com>