[omnibus] Don't create RPM database if it doesn't exist already
On Linux distributions not using the RPM package manager
(Debian, Ubuntu, etc.), the rpm probe of OpenSCAP could
initialize the sqlite RPM database into the /var/lib/rpm,
directory, as a side effect.

This change adds a new patch "rpmdb-no-create" to the RPM package.

This patch changes the openDatabase function to prevent creation
of the RPM database when it doesn't exist already.