regression detector: add compute merge base job
We need to compute the merge base of a non-`main` branch with respect
to `main` in order to establish the commit SHA of a baseline variant
in regression detection. This commit adds a job to compute that merge
base, echo the result to a file (sans newline), and then upload that
result to the Single-Machine Performance S3 bucket for Agent team.

Signed-off-by: Geoffrey M. Oxberry <geoffrey.oxberry@datadoghq.com>