regression detector: note build step we may need
The existing Single-Machine Performance (SMP) regression detector
setup doesn't build additional containers -- it just publishes
containers already built in Agent CI to SMP's ECR for Agent
images. I've written the new "merge-base" container assuming we can
continue with that strategy, but if we can't continue with that
strategy for some reason, then this commit includes a bunch of
comments sketching out a backup plan that would build the container we
would need and publish it to SMP's ECR for Agent images.

Signed-off-by: Geoffrey M. Oxberry <geoffrey.oxberry@datadoghq.com>