smp, docker_linux.yml: comment out merge base job
The merge base job isn't running because no runners are configured to
run it. This issue could be one that I can't resolve, so it could be
that I need to get permissions to run it. For now, I'll comment out
this job and instead try and get the information necessary to compute
a baseline as part of the regression detector job, but without
otherwise altering the regression detector behavior.

Signed-off-by: Geoffrey M. Oxberry <geoffrey.oxberry@datadoghq.com>