smp, docker_linux.yml: add container build notes
I've managed to figure out how to get the most important case to work
-- the one in which the base branch of the pull request (i.e., in
GitLab terms, the target branch of the merge request) is
`main`. Having managed to get this case to work, this commit updates
my implementation notes to reflect how to implement the case when the
base branch of the pull request *is not* `main`.

Signed-off-by: Geoffrey M. Oxberry <geoffrey.oxberry@datadoghq.com>