regression detector: set baseline to merge base
It looks like creating a separate job (at least in
`.gitlab/container_build/docker_linux.yml`) will not work because
GitLab won't run it in a runner. This behavior may be a configuration
setting for security reasons, and may require repo admin privileges to
change. Instead, this commit implements setting the baseline SHA to
the merge base, along with some commands to abort the regression
detector job if the base branch of the pull request (target branch of
the merge request) is not `main`, and copies the merge base SHA to S3
for debugging/auditing purposes.

Signed-off-by: Geoffrey M. Oxberry <geoffrey.oxberry@datadoghq.com>