Fix potentital use of uninitialized memory
This fixes potential use of uninitialized memory when PyList_GetItem
returns NULL.

This code path is impossible to hit in practice with the current
versions of Python, as long as the object is a list and index is in
bounds, which is ensured by the prior call to PyList_Size. These
functions do not use the Python sequence protocol, so evil python code
can not supply incorrect length or throw an unexpected exception
either.