regression detector: compute baseline via loop
The previous regression detector baseline SHA computation assumed that
an image exists in ECR for the commit returned by `$(git commit
merge-base HEAD main)`, *i.e.*, the merge base with `main`. However,
the container associated with that commit may not exist because that
container may have failed to build, or may have failed to upload to
ECR, so we need to check whether that container exists in ECR. If that
container exists, then the merge base with `main` is the baseline
SHA. If not, then we must iterate over predecessor commits in `main`
until we find a commit in `main` for which a container exists in
ECR. The first commit we find in this loop becomes the baseline SHA
for the regression detector.

This commit implements that check, along with a loop that iterates
over predecessor commits, if necessary, as described above. The
implementation is currently a rather verbose one-liner because
single-line statements are easier to debug in GitLab CI. Once this
implementation succeeds, a subsequent commit will clean up the
implementation by changing it to a multi-line statement, after which
this branch will be ready for review.

Signed-off-by: Geoffrey M. Oxberry <geoffrey.oxberry@datadoghq.com>