regression detector: reduce stderr noise
The initial implementation of the container existence check turns out
to be pretty noisy when a container doesn't exist because `awscli`
will output a bunch of error information on failure. While this
information is helpful for debugging purposes, it will be annoying in
CI, so this commit redirects the `stderr` of that command to
`/dev/null` to reduce noise in the CI output of the regression
detector CI job.

Signed-off-by: Geoffrey M. Oxberry <geoffrey.oxberry@datadoghq.com>