regression detector: tweak ecr query for debugging
For some reason, the `aws ecr describe-images` command is not finding
images that I know exist. To aid in debugging at the expense of noise,
this commit removes redirection of `stderr` to `/dev/null`, adds the
`--registry-id` flag to be more explicit about ECR repo location, and
also moves the `--profile` flag and its argument to a more readable
location.

Signed-off-by: Geoffrey M. Oxberry <geoffrey.oxberry@datadoghq.com>